linux-yocto-custom_3.10: Updated defconfig,SRCREV,patches
Update the defconfig to support hardware that is available for the
AT91SAM9G25 on the at91sam9x5ek evaluation board.  Add patches to
enable the RTC and USART3 in the device tree.  Add patches to fix
an RTC lockup bug that has been fixed in the mainline kernel.

Place patches and configuration files specific to the 3.10 kernel
in a directory based upon kernel version.

Remove patches that are no longer needed due to using an updated
SRCREV.

Signed-off-by: Bryan Evenson <bevenson@melinkcorp.com>