linux-yocto-custom_3.10: Remove reference to common defconfig
We use a specific defconfig that is located in a directory specific
to the target architecture and kernel version.  The base meta-atmel
recipe refers to a more generic defconfig that we do not use.  Remove
the reference to the generic defconfig so the build tool does not
complain about not being able to find it.

Signed-off-by: Bryan Evenson <bevenson@melinkcorp.com>